pyproject.toml
[project]
name = "budget-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.11.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
